process_workspace_deletion create and refactoring
