+++ studio/app/common/routers/experiment.py
@@ -82,7 +82,6 @@ async def delete_experiment(
 ):
     try:
         ExptDataWriter(workspace_id, unique_id).delete_data()
-
         if WorkspaceService.is_data_usage_available():
             WorkspaceService.delete_workspace_experiment(db, workspace_id, unique_id)
