+++ studio/app/common/routers/users_admin.py
@@ -89,48 +89,9 @@ async def delete_user(
         workspace_ids = [ws.id for ws in workspaces]
 
         for workspace_id in workspace_ids:
-            try:
-                # Delete experiment from database
-                WorkspaceService.delete_experiment_records_by_workspace_id(
-                    db, workspace_id
-                )
-
-                ws = (
-                    db.query(Workspace)
-                    .filter(
-                        Workspace.id == workspace_id,
-                        Workspace.user_id == user_id,
-                        Workspace.deleted.is_(False),
-                    )
-                    .first()
-                )
-
-                if not ws:
-                    raise HTTPException(
-                        status_code=404, detail=f"Workspace {workspace_id} not found"
-                    )
-
-                ws.deleted = True
-
-                db.add(ws)
-                db.commit()
-
-                WorkspaceService.delete_workspace_experiment_files(
-                    workspace_id=workspace_id, db=db
-                )
-
-            except Exception as e:
-                db.rollback()
-                logger.error(
-                    "Error deleting or updating workspace %s: %s",
-                    workspace_id,
-                    e,
-                    exc_info=True,
-                )
-                raise HTTPException(
-                    status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                    detail=f"Failed to delete or update workspace {workspace_id}.",
-                )
+            WorkspaceService.process_workspace_deletion(
+                db=db, workspace_id=workspace_id, user_id=user_id
+            )
 
         # Delete user and Firebase account
         try:
